# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProductCategory = typeof ProductCategory[keyof typeof ProductCategory];


export const ProductCategory = {
  trending: 'trending',
  gadgets: 'gadgets',
  gear: 'gear',
} as const;

export interface Product {
  id: string;
  name: string;
  description: string;
  price: string;
  /** @nullable */
  badge: string | null;
  rating: number;
  reviewCount: number;
  category: ProductCategory;
  imageUrl: string;
  active: boolean;
  createdAt: string;
  updatedAt: string;
}

export type ProductInputCategory = typeof ProductInputCategory[keyof typeof ProductInputCategory];


export const ProductInputCategory = {
  trending: 'trending',
  gadgets: 'gadgets',
  gear: 'gear',
} as const;

export interface ProductInput {
  name: string;
  description: string;
  price: string;
  /** @nullable */
  badge?: string | null;
  rating: number;
  reviewCount: number;
  category: ProductInputCategory;
  imageUrl: string;
  active: boolean;
}

export type ProductUpdateCategory = typeof ProductUpdateCategory[keyof typeof ProductUpdateCategory];


export const ProductUpdateCategory = {
  trending: 'trending',
  gadgets: 'gadgets',
  gear: 'gear',
} as const;

export interface ProductUpdate {
  name?: string;
  description?: string;
  price?: string;
  /** @nullable */
  badge?: string | null;
  rating?: number;
  reviewCount?: number;
  category?: ProductUpdateCategory;
  imageUrl?: string;
  active?: boolean;
}

export interface AdminLoginInput {
  username: string;
  password: string;
}

export interface AuthStatus {
  authenticated: boolean;
  /** @nullable */
  username: string | null;
}