# lib/api-zod/src/generated/types/authStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthStatus {
  authenticated: boolean;
  /** @nullable */
  username: string | null;
}