# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductCategory } from './productCategory';

export interface Product {
  id: string;
  name: string;
  description: string;
  price: string;
  /** @nullable */
  badge: string | null;
  rating: number;
  reviewCount: number;
  category: ProductCategory;
  imageUrl: string;
  active: boolean;
  createdAt: Date;
  updatedAt: Date;
}