# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductInputCategory } from './productInputCategory';

export interface ProductInput {
  name: string;
  description: string;
  price: string;
  /** @nullable */
  badge?: string | null;
  rating: number;
  reviewCount: number;
  category: ProductInputCategory;
  imageUrl: string;
  active: boolean;
}