# lib/api-zod/src/generated/types/productUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductUpdateCategory } from './productUpdateCategory';

export interface ProductUpdate {
  name?: string;
  description?: string;
  price?: string;
  /** @nullable */
  badge?: string | null;
  rating?: number;
  reviewCount?: number;
  category?: ProductUpdateCategory;
  imageUrl?: string;
  active?: boolean;
}